Created and updated recycler views
Created a second Recycler View that displays the Actors images and name. When clicked, it will navigate to second screen with actor name and image. The current plan is to add a second view on the bottom with the pictures of their popular movies.

Need to implement detail views for popular movies.